refactor: extract article lead+TOC rendering into _ArticleBodyMixin
Fifth decomposition cut (~285 lines). Moves the contiguous lead-with-TOC
block — _iter_article_h2 / _first_article_h2 / _advance_cut_to_second_h2 /
_is_disambig_lead / _lead_density / _lead_with_toc plus the _ARTICLE_H2_RE,
_DISAMBIG_LEAD_PHRASES, _LEAD_PREAMBLE_RE, _DUPLICATED_H1_RE and
_EMPTY_LEAD_DENSITY_THRESHOLD constants — into openzim_mcp/article_body.py
as _ArticleBodyMixin.

Cleanest cluster yet: the only external dep is zim_operations (declared
under TYPE_CHECKING); every other reference is cluster-internal. Pure code
movement, no behavior change.

simple_tools.py 4,247 → 3,966 (5,896 → 3,966 across five mixin cuts;
-1,930 lines / -33%). Full suite green (2623 passed); mypy/flake8/isort clean.
